fix: Don't bind to meow-keypad-state-keymap
If any key in this keymap is bound to something other than
`self-insert-command`, it is impossible to use any keybinding containing
that key, because the binding will be interrupted by said command. This
can put Meow in a state where it is impossible to exit Keypad state.

These bindings appear to have been a mistake as nothing like this is
present in any of the example `meow-setup` functions.

A trivial example of the issue - one cannot use the binding `C-x 0`.